bs/requirements.txt
# 基础依赖
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0
pyyaml>=6.0
Pillow>=8.3.0
scikeras>=0.8.0
scikit-learn>=1.0.0



# 音视频处理
ffmpeg-python>=0.2.0
librosa>=0.8.1
praat-parselmouth>=0.4.0
webrtcvad>=2.0.10

# 人脸处理
py-feat>=0.5.0
mediapipe>=0.8.9

# 预训练模型
transformers>=4.18.0
timm>=0.5.4

# 可视化和日志
tensorboard>=2.8.0
wandb>=0.12.0
opencv-python>=4.5.5

# 工具和辅助库
einops>=0.4.1
omegaconf>=2.1.1
hydra-core>=1.1.1
gradio>=4.0.0